# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.GenerateServiceTokenRequest:
    properties:
      api_key:
        type: string
      username:
        type: string
    required:
    - api_key
    - username
    type: object
  handlers.GenerateServiceTokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        properties:
          email:
            type: string
          id:
            type: integer
          role:
            type: string
          user_type:
            type: string
          username:
            type: string
        type: object
    type: object
...
  models.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  models.APIKeyCreateRequest:
    properties:
      allowed_domains:
        items:
          type: string
        type: array
      allowed_ips:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
      daily_quota:
        minimum: 1
        type: integer
      description:
        maxLength: 1000
        type: string
      expires_at:
        type: string
      metadata:
        $ref: '#/definitions/models.CustomFields'
      monthly_quota:
        minimum: 1
        type: integer
      name:
        maxLength: 255
        minLength: 1
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.APIPermissionType'
        minItems: 1
        type: array
      rate_limit_count:
        maximum: 100000
        minimum: 1
        type: integer
      rate_limit_window:
        allOf:
        - $ref: '#/definitions/models.RateLimitType'
        enum:
        - per_minute
        - per_hour
        - per_day
        - per_month
      scope_projects:
        items:
          type: integer
        type: array
      scope_users:
        items:
          type: integer
        type: array
      tags:
        items:
          type: string
        type: array
      user_agent_pattern:
        maxLength: 500
        type: string
    required:
    - name
    - permissions
    - rate_limit_window
    type: object
  models.APIKeyCreateResponse:
    properties:
      api_key:
        $ref: '#/definitions/models.APIKeyResponse'
      plain_key:
        description: Only returned once upon creation
        type: string
      plain_secret:
        description: Only returned once upon creation
        type: string
    type: object
...
  models.APIKeyResponse:
    properties:
      allowed_domains:
        items:
          type: string
        type: array
      allowed_ips:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
      created_at:
        type: string
      created_by:
        type: integer
      daily_quota:
        type: integer
      description:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      key_prefix:
        type: string
      last_used_at:
        type: string
      metadata:
        $ref: '#/definitions/models.CustomFields'
      monthly_quota:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.APIPermissionType'
        type: array
      rate_limit_count:
        type: integer
      rate_limit_window:
        $ref: '#/definitions/models.RateLimitType'
      scope_projects:
        items:
          type: integer
        type: array
      scope_users:
        items:
          type: integer
        type: array
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
      updated_by:
        type: integer
      usage_count:
        type: integer
      user_agent_pattern:
        type: string
    type: object
...
  models.APIPermissionType:
    enum:
    - api.read
    - api.write
    - api.admin
    - tasks.read
    - tasks.write
    - projects.read
    - projects.write
    - users.read
    - users.write
    - analytics.read
    - analytics.write
    - system.monitor
    type: string
    x-enum-varnames:
    - PermissionAPIRead
    - PermissionAPIWrite
    - PermissionAPIAdmin
    - PermissionTasksRead
    - PermissionTasksWrite
    - PermissionProjectsRead
    - PermissionProjectsWrite
    - PermissionUsersRead
    - PermissionUsersWrite
    - PermissionAnalyticsRead
    - PermissionAnalyticsWrite
    - PermissionSystemMonitor
...
  models.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.APIError'
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  models.CustomFields:
    additionalProperties: true
    type: object
...
  models.ErrorResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.APIError'
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  models.RateLimitType:
    enum:
    - per_minute
    - per_hour
    - per_day
    - per_month
    type: string
    x-enum-varnames:
    - RateLimitPerMinute
    - RateLimitPerHour
    - RateLimitPerDay
    - RateLimitPerMonth
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: AI项目管理后端API服务，提供项目管理、任务管理、文档生成等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: AI Project Backend API
  version: "1.0"
paths:
...
  /auth/service-api-key:
    post:
      consumes:
      - application/json
      description: Create a new API key for service account (admin only)
      parameters:
      - description: API Key creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.APIKeyCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: API Key created successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.APIKeyCreateResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create API key for service account
      tags:
      - ServiceAccount
  /auth/service-token:
    post:
      consumes:
      - application/json
      description: Generate JWT token pair for service account using API Key
      parameters:
      - description: API Key credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.GenerateServiceTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token generated successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.GenerateServiceTokenResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Generate service account token
      tags:
      - ServiceAccount
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"